GERONIMO-4732 click monitoring and delete the server localhost and there is an Error rendering portlet.(Patch from Rex Wang)
git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@798442 13f79535-47bb-0310-9956-ffa450edef68

+++ plugins/monitoring/mconsole-war/src/main/java/org/apache/geronimo/monitoring/console/MonitoringPortlet.java
@@ -679,7 +679,13 @@ public class MonitoringPortlet extends BasePortlet {
         try {
             userTransaction.begin();
             try {
-                Node node = entityManager.find(Node.class, server_id);
+            	Node node = entityManager.find(Node.class, server_id);
+            	// check if there is any graph created against the node before delete it.
+            	List<Graph> graphs = entityManager.createNamedQuery("graphsByNode").setParameter("name", node.getName()).getResultList();
+            	if (!(graphs == null || graphs.isEmpty())) {
+            		addErrorMessage(actionRequest, getLocalizedString(actionRequest, "mconsole.errorMsg20"));
+            		return;
+            	} 
                 entityManager.remove(node);
             } finally {
                 userTransaction.commit();
